removed serialization of basic types for member data fixed leak apparent when constant arrays are set in prepared objects rename pthreads_serial to pthreads_store everywhere to keep things organized

+++ src/globals.c
@@ -55,8 +55,8 @@ void pthreads_globals_unlock(int *acquired) {
 } /* }}} */
 
 /* {{{ pthreads_globals_count */
-long pthreads_globals_count() {
-	long result = 0L;
+size_t pthreads_globals_count() {
+	size_t result = 0L;
 	int locked;
 	
 	if (pthreads_globals_lock(&locked)) {
